SA-BE: Properties corrections into Siniestro model

+++ src/controllers/controlador.js
@@ -239,23 +239,35 @@ const updateControlador = async (req, res) => {
 const deleteControlador = async (req, res) => {
   const { id } = req.params;
   try {
-    const controladores = await Controladores.findByIdAndDelete(id);
-    if (controladores) {
-      res.status(200).json({
-        message: `Controlador ${id} deleted`,
-        data: controladores,
-        error: false,
-      });
-    } else {
-      res.status(404).json({
+    const existingControlador = await Controladores.findOne({ _id: id });
+
+    if (!existingControlador) {
+      return res.status(404).json({
         message: 'Controlador not found',
         data: null,
-        error: false,
+        error: true,
       });
     }
+    const { firebaseUid } = existingControlador;
+
+    await firebaseApp.auth().deleteUser(firebaseUid);
+
+    const result = await Controladores.findByIdAndDelete(id);
+    if (!result) {
+      return res.status(404).json({
+        message: 'Controlador not found',
+        data: null,
+        error: true,
+      });
+    }
+    return res.status(200).json({
+      message: 'Controlador deleted',
+      data: null,
+      error: false,
+    });
   } catch (error) {
-    res.status(500).json({
-      message: 'Internal Server Error',
+    return res.status(500).json({
+      message: error,
       data: null,
       error: true,
     });
